# docs/swagger.yaml
basePath: /
definitions:
...
  CreateRoomRequest:
    properties:
      capacity:
        example: 10
        type: integer
      description:
        example: Большая переговорная на 10 мест
        type: string
      name:
        example: Переговорная №1
        type: string
    type: object
  CreateRoomResponse:
    properties:
      room:
        $ref: '#/definitions/Room'
    type: object
  CreateScheduleResponse:
    properties:
      schedule:
        $ref: '#/definitions/Schedule'
    type: object
...
  ErrorDetail:
    properties:
      code:
        example: INVALID_REQUEST
        type: string
      message:
        example: invalid request
        type: string
    type: object
  ErrorResponse:
    properties:
      error:
        $ref: '#/definitions/ErrorDetail'
    type: object
...
  InternalErrorResponse:
    properties:
      error:
        $ref: '#/definitions/ErrorDetail'
    type: object
...
  ListRoomsResponse:
    properties:
      rooms:
        items:
          $ref: '#/definitions/Room'
        type: array
    type: object
  ListSlotsResponse:
    properties:
      slots:
        items:
          $ref: '#/definitions/Slot'
        type: array
    type: object
...
  Room:
    properties:
      capacity:
        type: integer
      createdAt:
        format: date-time
        type: string
      description:
        type: string
      id:
        format: uuid
        type: string
      name:
        type: string
    type: object
  Schedule:
    properties:
      daysOfWeek:
        items:
          type: integer
        type: array
      endTime:
        type: string
      id:
        format: uuid
        type: string
      roomId:
        format: uuid
        type: string
      startTime:
        type: string
    type: object
  Slot:
    properties:
      end:
        format: date-time
        type: string
      id:
        format: uuid
        type: string
      roomId:
        format: uuid
        type: string
      start:
        format: date-time
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: Сервис бронирования переговорок
  title: Room Booking Service
  version: 1.0.0
paths:
...
  /rooms/{roomId}/schedule/create:
    post:
      consumes:
      - application/json
      description: Доступно только роли admin. Расписание можно создать только один
        раз для каждой переговорки. Поле `daysOfWeek` должно содержать значения от
        1 до 7 (1=Пн, 7=Вс). При передаче значений вне этого диапазона возвращается
        400.
      parameters:
      - description: Идентификатор переговорки
        in: path
        name: roomId
        required: true
        type: string
      - description: Данные расписания
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/Schedule'
      produces:
      - application/json
      responses:
        "201":
          description: Расписание сохранено
          schema:
            $ref: '#/definitions/CreateScheduleResponse'
        "400":
          description: Неверный запрос (в т.ч. недопустимые значения daysOfWeek)
          schema:
            $ref: '#/definitions/ErrorResponse'
        "401":
          description: Не авторизован
          schema:
            $ref: '#/definitions/ErrorResponse'
        "403":
          description: Доступ запрещён (требуется роль admin)
          schema:
            $ref: '#/definitions/ErrorResponse'
        "404":
          description: Переговорка не найдена
          schema:
            $ref: '#/definitions/ErrorResponse'
        "409":
          description: Расписание для переговорки уже создано, изменение не допускается
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/InternalErrorResponse'
      security:
      - BearerAuth: []
      summary: Создать расписание переговорки (только admin, только один раз). Длительность
        слота 30 мин. После создания расписание изменить нельзя.
      tags:
      - Schedules
  /rooms/{roomId}/slots/list:
    get:
      description: Возвращает слоты, не занятые активной бронью, для указанной переговорки
        на указанную дату. Все даты и время передаются и возвращаются в UTC. Параметр
        `date` является обязательным; при его отсутствии возвращается 400. Если у
        переговорки нет расписания — возвращается пустой список (переговорка считается
        всегда недоступной).
      parameters:
      - description: Идентификатор переговорки
        in: path
        name: roomId
        required: true
        type: string
      - description: 'Дата в формате ISO 8601 (например: 2024-06-10). Обязательный
          параметр; при отсутствии — 400.'
        in: query
        name: date
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Список доступных слотов (не занятых активной бронью). Пустой
            список, если у переговорки нет расписания или на эту дату нет слотов.
          schema:
            $ref: '#/definitions/ListSlotsResponse'
        "400":
          description: Неверный запрос (отсутствует или некорректен параметр date)
          schema:
            $ref: '#/definitions/ErrorResponse'
        "401":
          description: Не авторизован
          schema:
            $ref: '#/definitions/ErrorResponse'
        "404":
          description: Переговорка не найдена
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/InternalErrorResponse'
      security:
      - BearerAuth: []
      summary: Список доступных для бронирования слотов по переговорке и дате (admin
        и user). Наиболее нагруженный эндпоинт.
      tags:
      - Slots
  /rooms/create:
    post:
      consumes:
      - application/json
      parameters:
      - description: Данные переговорки
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/CreateRoomRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Переговорка создана
          schema:
            $ref: '#/definitions/CreateRoomResponse'
        "400":
          description: Неверный запрос
          schema:
            $ref: '#/definitions/ErrorResponse'
        "401":
          description: Не авторизован
          schema:
            $ref: '#/definitions/ErrorResponse'
        "403":
          description: Доступ запрещён (требуется роль admin)
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/InternalErrorResponse'
      security:
      - BearerAuth: []
      summary: Создать переговорку (только admin)
      tags:
      - Rooms
  /rooms/list:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: Список переговорок
          schema:
            $ref: '#/definitions/ListRoomsResponse'
        "401":
          description: Не авторизован
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/InternalErrorResponse'
      security:
      - BearerAuth: []
      summary: Список переговорок (admin и user)
      tags:
      - Rooms
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
...